chunking/requirements.txt
# Core document processing (using newer versions compatible with Python 3.13)
unstructured>=0.16.0
pypdf>=4.0.0
python-docx>=1.1.0
markdown>=3.6
beautifulsoup4>=4.12.3
lxml>=5.0.0

# OpenAI embeddings (newer version)
openai>=1.54.0

# Database (using psycopg3 - pure Python implementation)
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0

# Configuration and utilities
python-dotenv>=1.0.0
PyYAML>=6.0.1
tenacity>=8.2.3
tqdm>=4.66.1

# Logging and validation (use newer version with pre-built wheels for Python 3.13)
pydantic>=2.9.0
